标签功能-继续改

+++ src/main/java/com/book/domain/BookReservation.java
@@ -5,9 +5,9 @@ import java.util.Date;
 
 @Data
 public class BookReservation {
-    private Integer id;         // 预约ID
+    private Long id;         // 预约ID
     private Long bookId;        // 图书ID
-    private Integer readerId;   // 读者ID
+    private Long readerId;   // 读者ID
     private Date reserveTime;   // 预约时间
     private Integer status;     // 状态：0=等待中, 1=已通知, 2=已取消
     private Date notifyTime;    // 通知时间
