+++ src/main/java/com/book/mapper/NotificationMapper.java
@@ -13,19 +13,19 @@ public interface NotificationMapper {
     void insert(Notification notification);
     
     @Select("SELECT * FROM notification WHERE user_id = #{userId} AND status = 0 ORDER BY create_time DESC")
-    List<Notification> findUnreadByUserId(Integer userId);
+    List<Notification> findUnreadByUserId(Long userId);
     
     @Select("SELECT * FROM notification WHERE user_id = #{userId} ORDER BY create_time DESC")
-    List<Notification> findByUserId(Integer userId);
+    List<Notification> findByUserId(Long userId);
     
     @Update("UPDATE notification SET status = #{status}, read_time = NOW() WHERE notification_id = #{notificationId}")
     int updateStatus(@Param("notificationId") Long notificationId, @Param("status") Integer status);
     
     @Update("UPDATE notification SET status = 1, read_time = NOW() WHERE user_id = #{userId} AND status = 0")
-    int markAllAsRead(Integer userId);
+    int markAllAsRead(Long userId);
 
     @Select("SELECT COUNT(*) FROM notification WHERE user_id = #{userId} AND status = 0")
-    int countUnreadByUserId(Integer userId);
+    int countUnreadByUserId(Long userId);
 
     @Update("UPDATE notification SET status = 1, read_time = NOW() WHERE notification_id = #{notificationId}")
     void markAsRead(Long notificationId);
